// src/core/types/album/Face.ts
export interface Face {
  d_width: number;
  d_height: number;
  d_top: number;
  d_left: number;

  width: number;
  height: number;
  top: number;
  left: number;

  scaledWidth: number;
  scaledHeight: number;
  offsetX: number;
  offsetY: number;
  resizedPhotoWidth: number;
  resizedPhotoHeight: number;
}